backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
sqlalchemy==2.0.4
psycopg2-binary==2.9.5
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
fpdf==1.7.2
gunicorn==20.1.0
pydantic==1.10.7
email-validator==2.0.0
bcrypt==4.0.1
pandas==1.5.3
openpyxl==3.1.2
xlrd==2.0.1

